Update dev.js file
Include fast inicialization option

+++ dev.js
@@ -5,6 +5,8 @@ import { authConfig } from './app/config/auth.config.js';
 
 const app = express();
 
+const fast = true;
+
 // Configuração do aplicativo
 app.use(cors({ origin: authConfig.corsOrigin }));
 app.use(json());
@@ -37,23 +39,29 @@ import addData from './app/data/add.data.js';
 // Conexão com o banco de dados
 conectarAoBanco();
 
-// In development or in first time run, use this code
-// :::::::: Caution ::::::::
-// This code will drop all tables everytime
+// In development use this code
 async function conectarAoBanco() {
     try {
         await db.sequelize.authenticate();
         console.log("Conexão estabelicida com sucesso.");
 
-        await db.sequelize.sync({ force: true }).then(async () => {
-            console.log('Removendo e Resincronizando o Banco de Dados');
+        if (fast) {
+            //For fast inicialization
+            await db.sequelize.sync();
+        } else {
+            // :::::::: Caution ::::::::
+            // This code will drop all tables everytime
+            await db.sequelize.sync({ force: true }).then(async () => {
+                console.log('Removendo e Resincronizando o Banco de Dados');
 
-            await addData.cargo();
-            await addData.estado();
-            await addData.cidade();
-            await addData.categoria();
-            await addData.usuario();
-        });
+                await addData.cargo();
+                await addData.estado();
+                await addData.cidade();
+                await addData.categoria();
+                await addData.usuario();
+            });
+        }
+
         console.log("Todos os modelos foram sincronizados com sucesso.");
     } catch (error) {
         console.error("Não foi possível se conectar ao banco de dados:", error);
